Add 32-bit funcs and document front end (#39)
* 32-bit functions
* adapt tests since the default write/read_num is now 32-bit
* minor cleaning in the generator
* add test for _32 functionality
* rm stdint-based types from test
* documentation
* rm trailing whitespaces in emacs
* change TREXIO_INVALID_BACK_END
* update gitignore

+++ src/test.f90
@@ -15,7 +15,7 @@ subroutine test_write()
   integer(8) :: trex_file
 
   integer :: rc = 1
-  integer(8) :: num
+  integer :: num
 
   double precision :: charge(12)
   double precision :: coord(36)
@@ -94,7 +94,7 @@ subroutine test_read()
   integer(8) :: trex_file
 
   integer :: rc = 1
-  integer(8) :: num, num_read
+  integer :: num, num_read
 
   double precision :: charge(12)
   double precision :: coord(3,12)
